+++ src/linker/Linker.Steps/ValidateVirtualMethodAnnotationsStep.cs
@@ -45,7 +45,7 @@ namespace Mono.Linker.Steps
 				string message = MessageFormat.FormatRequiresAttributeMismatch (methodHasAttribute,
 					baseMethod.DeclaringType.IsInterface, nameof (RequiresUnreferencedCodeAttribute), method.GetDisplayName (), baseMethod.GetDisplayName ());
 				Context.LogWarning (new DiagnosticString (DiagnosticId.RequiresUnreferencedCodeAttributeMismatch).GetMessage (message),
-					(int) DiagnosticId.RequiresUnreferencedCodeAttributeMismatch, method, subcategory: MessageSubCategory.TrimAnalysis);
+					DiagnosticId.RequiresUnreferencedCodeAttributeMismatch, method, subcategory: MessageSubCategory.TrimAnalysis);
 			}
 		}
 	}
